+++ web/frontend/output/dev/server/server-reference-manifest.js
@@ -1,0 +1,1 @@
+self.__RSC_SERVER_MANIFEST="{\n  \"node\": {},\n  \"edge\": {},\n  \"encryptionKey\": \"8LWcYS51aq+vItAulW/I2P8GVK0DBQS/3Rpq71lteSQ=\"\n}"