+++ web/frontend/output/dev/static/chunks/app_layout_tsx_1cf6b850._.js
@@ -1,0 +1,9 @@
+(globalThis.TURBOPACK_CHUNK_LISTS || (globalThis.TURBOPACK_CHUNK_LISTS = [])).push({
+    script: typeof document === "object" ? document.currentScript : undefined,
+    chunks: [
+  "static/chunks/[root-of-the-server]__2769d28c._.css",
+  "static/chunks/_56487040._.js",
+  "static/chunks/node_modules_f8897084._.js"
+],
+    source: "dynamic"
+});